+++ src/pages/AdminMainPage/Components/ApplicantDetail.js
